+++ Assets/Scripts/Model/EditData.cs
@@ -12,7 +12,7 @@ namespace NoteEditor.Model
         static ReactiveProperty<string> name_ = new ReactiveProperty<string>();
         static ReactiveProperty<int> maxBlock_ = new ReactiveProperty<int>(5);
         static ReactiveProperty<int> LPB_ = new ReactiveProperty<int>(4);
-        static ReactiveProperty<int> BPM_ = new ReactiveProperty<int>(120);
+        static ReactiveProperty<float> BPM_ = new ReactiveProperty<float>(120f);
         static ReactiveProperty<int> offsetSamples_ = new ReactiveProperty<int>(0);
         static Dictionary<NotePosition, NoteObject> notes_ = new Dictionary<NotePosition, NoteObject>();
 
@@ -20,7 +20,7 @@ namespace NoteEditor.Model
         public static ReactiveProperty<string> Name { get { return name_; } }
         public static ReactiveProperty<int> MaxBlock { get { return maxBlock_; } }
         public static ReactiveProperty<int> LPB { get { return LPB_; } }
-        public static ReactiveProperty<int> BPM { get { return BPM_; } }
+        public static ReactiveProperty<float> BPM { get { return BPM_; } }
         public static ReactiveProperty<int> OffsetSamples { get { return offsetSamples_; } }
         public static Dictionary<NotePosition, NoteObject> Notes { get { return notes_; } }
     }
